requirements.txt
pydantic~=1.10.6
graphene~=3.2.2
fastapi~=0.95.0
uvicorn~=0.21.1
strawberry-graphql~=0.165.0